move notebook views inside notebook module
