+++ frontend/src/modules/Notebooks/views/NotebookInstanceProperties.js
