+++ frontend/src/modules/Notebooks/views/NotebookList.js
@@ -18,9 +18,9 @@ import {
   PlusIcon,
   SearchInput,
   useSettings
-} from '../../../../design';
-import { SET_ERROR, useDispatch } from '../../../../globalErrors';
-import { listSagemakerNotebooks, useClient } from '../../../../services';
+} from '../../../design';
+import { SET_ERROR, useDispatch } from '../../../globalErrors';
+import { listSagemakerNotebooks, useClient } from '../../../services';
 import NotebookListItem from './NotebookListItem';
 
 function NotebookPageHeader() {
