+++ frontend/src/modules/Notebooks/views/NotebookListItem.js
@@ -15,12 +15,7 @@ import * as FiIcons from 'react-icons/fi';
 import { SiJupyter } from 'react-icons/si';
 import { useNavigate } from 'react-router';
 import { Link as RouterLink } from 'react-router-dom';
-import {
-  IconAvatar,
-  Label,
-  StackStatus,
-  useCardStyle
-} from '../../../../design';
+import { IconAvatar, Label, StackStatus, useCardStyle } from '../../../design';
 
 /**
  * @description NotebookListItem view.
